+++ product/src/main/java/vn/whatsenglish/product/enums/Discounts.java
@@ -7,15 +7,15 @@ import vn.whatsenglish.product.strategy.discount.PercentageDiscountStrategy;
 
 public enum Discounts {
 
-    FIXED_DISCOUNT(1, FixedDiscountStrategy.class),
-    PERCENTAGE_DISCOUNT(2, PercentageDiscountStrategy.class);
+    FIXED_DISCOUNT(1L, FixedDiscountStrategy.class),
+    PERCENTAGE_DISCOUNT(2L, PercentageDiscountStrategy.class);
 
     @Getter
-    private final Integer id;
+    private final Long id;
     @Getter
     private final Class<? extends IDiscountStrategy> discountStrategy;
 
-    Discounts(Integer id, Class<? extends IDiscountStrategy> discountStrategy) {
+    Discounts(Long id, Class<? extends IDiscountStrategy> discountStrategy) {
         this.id = id;
         this.discountStrategy = discountStrategy;
     }
